Feat. BettterGrowthMomentTraits
Add scores to some traits. Higher score means that it's more likely to be in child's growth moment traits if the child has high enough growth tier, and vice versa.
This should help to reduce the frustration when trying to raise a child.
(Note: this doesn't mean there will be no negative traits when at the highest growth tier)